requirements.txt
absl-py==2.1.0
accelerate==0.30.0
annotated-types==0.7.0
asttokens==2.4.1
backcall==0.2.0
blessed==1.20.0
cachetools==5.3.3
calmsize==0.1.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.8
cmake==3.25.0
colorama==0.4.6
comm==0.2.2
contourpy==1.1.1
cycler==0.12.1
DateTime==5.5
debugpy==1.8.1
decorator==5.1.1
docker-pycreds==0.4.0
easyocr==1.7.1
einops==0.7.0
einops-exts==0.0.4
einx==0.1.3
ema-pytorch==0.7.7
eval_type_backport==0.2.2
executing==2.0.1
filelock==3.13.3
fonttools==4.51.0
frozendict==2.4.1
fsspec==2024.3.1
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.29.0
google-auth-oauthlib==0.4.6
gpustat==1.1.1
graphviz==0.20.3
grpcio==1.62.1
huggingface-hub==0.22.2
idna==3.6
imageio==2.35.1
importlib_metadata==7.1.0
importlib_resources==6.4.0
ipykernel==6.29.4
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.3
jiwer==3.0.5
joblib==1.4.2
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
lazy_loader==0.4
lit==15.0.7
lmdb==1.5.1
lpips==0.1.4
lxml==5.2.1
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.6
memory-profiler==0.61.0
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.1
ninja==1.11.1.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.133
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python-headless==4.10.0.84
packaging==24.0
pandas==2.0.3
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
platformdirs==4.2.0
portalocker==2.8.2
prompt-toolkit==3.0.43
protobuf==5.26.1
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyclipper==1.3.0.post5
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.17.2
pyparsing==3.1.2
python-bidi==0.6.0
python-dateutil==2.9.0.post0
pytorch-fid==0.3.0
pytorch-memlab==0.3.0
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.2
rapidfuzz==3.9.7
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
sacrebleu==2.4.2
safetensors==0.4.2
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
sentencepiece==0.2.0
sentry-sdk==2.19.2
setproctitle==1.3.4
shapely==2.0.6
six==1.16.0
smmap==5.0.2
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tensorboard==2.12.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==0.9.16
tokenizers==0.15.2
torch==2.0.1+cu118
torch-summary==1.4.5
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
torchviz==0.0.3
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
transformers==4.39.2
triton==2.0.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.1
vector_quantize_pytorch==1.14.6
wandb==0.19.1
wcwidth==0.2.13
Werkzeug==3.0.2
zipp==3.18.1
zope.interface==6.2
